+++ src/main/resources/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 - Update build environment to minecraft Forge 10.13.2.1291
+- Added bat and squid spawn control.
 
 ## 1.1.0
 - Added clay generation to overworld.